Refactor apps for shame and voice journals
