+++ backend/core/models.py
@@ -2,24 +2,9 @@ from django.db import models
 from django.contrib.auth.models import User
 
 
-class Badge(models.Model):
-    """Awarded to users when they hit certain milestones."""
-
-    code = models.CharField(max_length=50, unique=True)
-    name = models.CharField(max_length=100)
-    description = models.TextField()
-    emoji = models.CharField(max_length=10)
-    is_active = models.BooleanField(default=True)
-
-    class Meta:
-        ordering = ["code"]
-
-    def __str__(self) -> str:  # pragma: no cover - simple representation
-        return self.code
-
-
 class Profile(models.Model):
     """User profile linked to the Django auth user."""
+
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     display_name = models.CharField(max_length=64)
     bio = models.TextField(blank=True)
@@ -27,48 +12,15 @@ class Profile(models.Model):
     last_active = models.DateTimeField(null=True, blank=True)
     current_mood = models.CharField(max_length=20, default="neutral")
     mood_last_updated = models.DateTimeField(null=True, blank=True)
-    badges = models.ManyToManyField(
-        Badge, related_name="owners", blank=True
-    )
+    badges = models.ManyToManyField("shame.Badge", related_name="owners", blank=True)
 
     def __str__(self) -> str:  # pragma: no cover - simple representation
         return self.display_name
 
 
-class DailyLockout(models.Model):
-    """Tracks whether the user completed required activity for the day."""
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    date = models.DateField()
-    is_unlocked = models.BooleanField(default=False)
-    required_minutes = models.IntegerField(default=30)
-    minutes_completed = models.IntegerField(default=0)
-
-    class Meta:
-        unique_together = ("user", "date")
-        ordering = ["-date"]
-
-    def __str__(self) -> str:  # pragma: no cover
-        return f"{self.user.username} - {self.date}"
-
-
-class ShamePost(models.Model):
-    """Donkey meme posted when a user misses a requirement."""
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    date = models.DateField()
-    image_url = models.URLField()
-    caption = models.TextField()
-    posted_to = models.JSONField(default=list)
-    was_triggered = models.BooleanField(default=False)
-
-    class Meta:
-        ordering = ["-date"]
-
-    def __str__(self) -> str:  # pragma: no cover
-        return f"ShamePost {self.user.username} {self.date}"
-
-
 class PaddleLog(models.Model):
     """Log of paddle sessions with optional mood and photo."""
+
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     date = models.DateField()
     location = models.CharField(max_length=128, blank=True)
@@ -83,59 +35,9 @@ class PaddleLog(models.Model):
         return f"PaddleLog {self.user.username} {self.date}"
 
 
-class VoiceJournal(models.Model):
-    """Uploaded audio journal with transcript and summary."""
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    audio_file = models.FileField(upload_to="voice_journals/")
-    transcript = models.TextField(blank=True)
-    summary = models.TextField(blank=True)
-    playback_audio_url = models.URLField(blank=True, null=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    tags = models.JSONField(blank=True, null=True)
-
-    class Meta:
-        ordering = ["-created_at"]
-
-    def __str__(self) -> str:  # pragma: no cover
-        return f"VoiceJournal {self.user.username} {self.created_at}"
-
-
-class Herd(models.Model):
-    """Group of users that share memes and badge shoutouts."""
-    name = models.CharField(max_length=100)
-    created_by = models.ForeignKey(
-        User, on_delete=models.CASCADE, related_name="created_herds"
-    )
-    members = models.ManyToManyField(User, related_name="herds")
-    tone = models.CharField(
-        max_length=20,
-        choices=[("roast", "Roast"), ("encourage", "Encourage"), ("mixed", "Mixed")],
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    invite_code = models.CharField(max_length=12, unique=True)
-
-    def __str__(self) -> str:  # pragma: no cover - simple representation
-        return self.name
-
-
-class BadgeShoutout(models.Model):
-    """Record a badge shoutout that can be shared with the user's herd."""
-
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    badge = models.ForeignKey(Badge, on_delete=models.CASCADE)
-    herd = models.ForeignKey(Herd, on_delete=models.CASCADE, null=True, blank=True)
-    message = models.TextField()
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        ordering = ["-created_at"]
-
-    def __str__(self) -> str:  # pragma: no cover - simple representation
-        return f"{self.user.username} - {self.badge.code}"
-
-
 class WorkoutLog(models.Model):
     """Record of a user's workout session details."""
+
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     activity_type = models.CharField(max_length=50)
     duration_minutes = models.IntegerField()
@@ -171,51 +73,14 @@ class MovementGoal(models.Model):
         ordering = ["-created_at"]
 
     def __str__(self) -> str:  # pragma: no cover - simple representation
-        return f"{self.user.username} {self.activity_type} {self.target_sessions}" \
+        return (
+            f"{self.user.username} {self.activity_type} {self.target_sessions}"
             f" {self.start_date}->{self.end_date}"
-
-
-class DonkeyChallenge(models.Model):
-    """Short challenge issued to motivate the user after failures."""
-
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    challenge_text = models.TextField()
-    issued_at = models.DateTimeField(auto_now_add=True)
-    expires_at = models.DateTimeField()
-    is_completed = models.BooleanField(default=False)
-    is_failed = models.BooleanField(default=False)
-    tone = models.CharField(max_length=20, default="mixed")
-    auto_generated = models.BooleanField(default=True)
-
-    class Meta:
-        ordering = ["-issued_at"]
-
-    def __str__(self) -> str:  # pragma: no cover - simple representation
-        return f"{self.user.username} - {self.tone}"
-
-
-class HerdPost(models.Model):
-    """Simple post shared with the user's herd."""
-
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    type = models.CharField(max_length=20)  # meme or badge
-    caption = models.TextField(blank=True)
-    image_url = models.URLField(blank=True, null=True)
-    emoji = models.CharField(max_length=5, blank=True)
-    badge_name = models.CharField(max_length=100, blank=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        ordering = ["-created_at"]
-
-    def __str__(self) -> str:  # pragma: no cover - simple representation
-        return f"{self.user.username} {self.type} {self.created_at}"
+        )
 
 
 class DailyGoal(models.Model):
-
     """Simple per-day goal entry."""
-
 
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     goal = models.CharField(max_length=50)
@@ -231,4 +96,3 @@ class DailyGoal(models.Model):
 
     def __str__(self) -> str:  # pragma: no cover - simple representation
         return f"{self.user.username} {self.goal} {self.date}"
-
