+++ backend/server/settings.py
@@ -47,8 +47,9 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "rest_framework",
     "rest_framework.authtoken",
-
     "core",
+    "shame",
+    "voice_journals",
     "prompts",
     "movement",
     "content",
