+++ backend/voice_journals/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import VoiceJournal
+
+
+@admin.register(VoiceJournal)
+class VoiceJournalAdmin(admin.ModelAdmin):
+    list_display = ("user", "created_at")
