+++ backend/voice_journals/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .views import upload_voice_journal
+
+urlpatterns = [
+    path("upload/", upload_voice_journal),
+]
