+++ backend/voice_journals/utils/tts_helpers.py
@@ -4,19 +4,14 @@ import os
 ELEVENLABS_API_KEY = os.getenv("ELEVENLABS_API_KEY")
 VOICE_ID = "EXAVITQu4vr4xnSDxMaL"  # Replace with preferred stock voice
 
+
 def text_to_speech(text, output_path="output.mp3"):
     url = f"https://api.elevenlabs.io/v1/text-to-speech/{VOICE_ID}"
-    headers = {
-        "xi-api-key": ELEVENLABS_API_KEY,
-        "Content-Type": "application/json"
-    }
+    headers = {"xi-api-key": ELEVENLABS_API_KEY, "Content-Type": "application/json"}
     payload = {
         "text": text,
         "model_id": "eleven_monolingual_v1",
-        "voice_settings": {
-            "stability": 0.4,
-            "similarity_boost": 0.6
-        }
+        "voice_settings": {"stability": 0.4, "similarity_boost": 0.6},
     }
     response = requests.post(url, headers=headers, json=payload)
     if response.status_code == 200:
